Add fishery package & move the driver card out as a single component
