+++ web-client/package.json
@@ -25,6 +25,7 @@
     "@material-ui/core": "^4.11.2",
     "@reduxjs/toolkit": "^1.5.0",
     "@types/react-redux": "^7.1.15",
+    "fishery": "^1.2.0",
     "next": "10.0.3",
     "react": "17.0.1",
     "react-dom": "17.0.1",
