+++ web-client/src/shared/test-utils/factories.ts
@@ -1,10 +1,11 @@
+import { Factory } from 'fishery';
 import { Driver } from '../redux/driversSearch/index';
 
 // mock driver information
-export const createDriverInfo = (id: number, name = 'Pam Lutz', email?: string): Driver => ({
-  id,
-  name: name,
-  email: email || '',
+export const DriverFactory = Factory.define<Driver>(({ sequence }) => ({
+  id: sequence,
+  name: 'Pam Lutz',
+  email: '',
   licenceClass: 'A-Z',
   vehicleType: 'Truck',
   policeCheck: 'Last 12 months',
@@ -12,4 +13,4 @@ export const createDriverInfo = (id: number, name = 'Pam Lutz', email?: string):
   willingToLift: 'Up to 50lbs',
   packingAndSorting: 'Yes',
   riskComfortLevel: ['Contactless Deliveries', 'Low Risk', 'High Risk'],
-})
+}));
